+++ config/routes.js
@@ -1,0 +1,5 @@
+var routes = [
+  { path: '/', name: 'index', controller: 'home' }
+];
+
+module.exports = routes;